+++ packages/@tinacms/toolkit/src/packages/fields/components/Reference.tsx
@@ -21,6 +21,7 @@ import styled from 'styled-components'
 import { BiEdit } from 'react-icons/bi'
 import { useCMS } from '../../../react-tinacms/use-cms'
 import { TinaCMS } from '../../../tina-cms'
+import { selectFieldClasses } from './Select'
 
 type Option = {
   value: string
@@ -94,11 +95,12 @@ export const Reference: React.FC<ReferenceProps> = ({
 
   const selectOptions = options || field.options
   return (
-    <SelectElement>
+    <div>
       <select
         id={input.name}
         value={input.value}
         onChange={input.onChange}
+        className={selectFieldClasses}
         {...input}
       >
         {selectOptions ? (
@@ -114,31 +116,15 @@ export const Reference: React.FC<ReferenceProps> = ({
               href={`/admin/collections/${
                 document.sys.collection.name
               }/${document.sys.breadcrumbs.join('/')}`}
-              style={{
-                display: 'flex',
-                alignItems: 'center',
-                textTransform: 'uppercase',
-                fontSize: '12px',
-                fontWeight: 500,
-                marginTop: '8px',
-                marginBottom: '4px',
-                lineHeight: '1',
-              }}
+              className="text-gray-700 hover:text-blue-500 flex items-center uppercase text-sm mt-2 mb-2 leading-none"
             >
-              <BiEdit
-                style={{
-                  height: '1.365em',
-                  width: 'auto',
-                  opacity: '0.8',
-                  marginRight: '6px',
-                }}
-              />
+              <BiEdit className="h-5 w-auto opacity-80 mr-2" />
               Edit in Tina Admin
             </a>
           )}
         </GetReference>
       )}
-    </SelectElement>
+    </div>
   )
 }
 function toProps(option: Option | string): Option {
@@ -153,52 +139,3 @@ function toComponent(option: Option) {
     </option>
   )
 }
-
-const SelectElement = styled.div`
-  display: block;
-  position: relative;
-
-  select {
-    all: unset;
-    box-sizing: border-box;
-    display: block;
-    font-family: inherit;
-    max-width: 100%;
-    color: var(--tina-color-grey-10);
-    padding: var(--tina-padding-small);
-    border-radius: var(--tina-radius-small);
-    background: var(--tina-color-grey-0);
-    font-size: var(--tina-font-size-2);
-    font-weight: var(--tina-font-weight-regular);
-    line-height: 1.35;
-    position: relative;
-    background-color: var(--tina-color-grey-0);
-    transition: all 85ms ease-out;
-    border: 1px solid var(--tina-color-grey-2);
-    width: 100%;
-    margin: 0;
-    appearance: none;
-    outline: none;
-    background-image: url('data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%22292.4%22%20height%3D%22292.4%22%3E%3Cpath%20fill%3D%22%23007CB2%22%20d%3D%22M287%2069.4a17.6%2017.6%200%200%200-13-5.4H18.4c-5%200-9.3%201.8-12.9%205.4A17.6%2017.6%200%200%200%200%2082.2c0%205%201.8%209.3%205.4%2012.9l128%20127.9c3.6%203.6%207.8%205.4%2012.8%205.4s9.2-1.8%2012.8-5.4L287%2095c3.5-3.5%205.4-7.8%205.4-12.8%200-5-1.9-9.2-5.5-12.8z%22%2F%3E%3C%2Fsvg%3E');
-    background-repeat: no-repeat;
-    background-position: right 0.7em top 50%;
-    background-size: 0.65em auto;
-    padding-right: 1.5rem;
-    overflow: hidden;
-    text-overflow: ellipsis;
-
-    &:hover {
-      box-shadow: 0 0 0 2px var(--tina-color-grey-3);
-    }
-
-    &:focus {
-      box-shadow: 0 0 0 2px var(--tina-color-primary);
-    }
-  }
-
-  select:-moz-focusring,
-  select::-moz-focus-inner {
-    color: transparent;
-    text-shadow: 0 0 0 #000;
-  }
-`