+++ packages/@tinacms/toolkit/src/packages/icons/AlignCenter.tsx
@@ -18,8 +18,8 @@ limitations under the License.
 
 import * as React from 'react'
 
-export const AlignCenter = () => (
-  <svg viewBox="0 0 32 32" xmlns="http://www.w3.org/2000/svg">
+export const AlignCenter = ({ ...props }) => (
+  <svg viewBox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" {...props}>
     <path
       d="M9.125 24H22.875V26H9.125V24ZM5 18H27V20H5V18ZM5 6H27V8H5V6ZM9.125 12H22.875V14H9.125V12Z"
       fill="inherit"