+++ packages/next-tinacms-cloudinary/CHANGELOG.md
@@ -1,5 +1,9 @@
 # next-tinacms-cloudinary
 
+## 3.5.2
+
+## 3.5.1
+
 ## 3.5.0
 
 ### Minor Changes